+++ src/main/java/com/angelobdev/openindustry/OpenIndustry.java
@@ -1,14 +1,13 @@
 package com.angelobdev.openindustry;
 
-import com.angelobdev.openindustry.block.BlocksHolder;
-import com.angelobdev.openindustry.item.ItemsHolder;
+import com.angelobdev.openindustry.block.OIBlocks;
+import com.angelobdev.openindustry.item.OIItems;
 import com.angelobdev.openindustry.world.feature.OIConfiguredFeatures;
 import com.angelobdev.openindustry.world.feature.OIPlacedFeatures;
 import com.mojang.logging.LogUtils;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.slf4j.Logger;
@@ -27,14 +26,14 @@ public class OpenIndustry {
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
 
         modEventBus.addListener(this::commonSetup);
-        modEventBus.addListener(this::clientSetup);
 
         // OpenIndustry custom registries
-        BlocksHolder.register(modEventBus); // Blocks container
-        ItemsHolder.register(modEventBus); // Items container
+        OIBlocks.register(modEventBus); // Blocks container
+        OIItems.register(modEventBus); // Items container
 
+        // OI Features registries
         OIConfiguredFeatures.register(modEventBus);
-        OIPlacedFeatures.register(modEventBus); // Ore Spawning
+        OIPlacedFeatures.register(modEventBus);
 
         // Register ourselves for server and other game events we are interested in
         MinecraftForge.EVENT_BUS.register(this);
@@ -42,11 +41,6 @@ public class OpenIndustry {
 
     private void commonSetup(final FMLCommonSetupEvent event) {
         LOGGER.info("Welcome to OpenIndustry for Minecraft 1.19.2!");
-    }
-
-    private void clientSetup(final FMLClientSetupEvent event) {
-        //ItemBlockRenderTypes.setRenderLayer(BlocksHolder.RUBBER_LEAVES_BLOCK.get(), RenderType.cutout());
-        //ItemBlockRenderTypes.setRenderLayer(BlocksHolder.RUBBER_SAPLING.get(), RenderType.cutout());
     }
 
 }