+++ src/main/java/com/angelobdev/openindustry/block/OIBlocks.java
@@ -1,7 +1,7 @@
 package com.angelobdev.openindustry.block;
 
 import com.angelobdev.openindustry.OpenIndustry;
-import com.angelobdev.openindustry.item.ItemsHolder;
+import com.angelobdev.openindustry.item.OIItems;
 import com.angelobdev.openindustry.item.custom.OIFlammableRotatedPillarBlock;
 import com.angelobdev.openindustry.world.feature.tree.RubberTreeGrower;
 import net.minecraft.core.BlockPos;
@@ -9,6 +9,8 @@ import net.minecraft.core.Direction;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.LeavesBlock;
@@ -20,20 +22,21 @@ import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Supplier;
 
-import static com.angelobdev.openindustry.item.ItemsHolder.OI_TAB;
+import static com.angelobdev.openindustry.item.OIItems.OI_TAB;
 import static net.minecraft.world.level.block.Blocks.*;
 
 @SuppressWarnings("unused")
-public final class BlocksHolder {
+public final class OIBlocks {
 
     // Blocks Registry
     public static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, OpenIndustry.MODID);
 
     // region COPPER
-    public static final RegistryObject<Block> COPPER_ORE = BlocksHolder.registerBlock(
+    public static final RegistryObject<Block> COPPER_ORE = OIBlocks.registerBlock(
             "copper_ore",
             () -> new Block(
                     BlockBehaviour.Properties
@@ -44,7 +47,7 @@ public final class BlocksHolder {
             OI_TAB
     );
 
-    public static final RegistryObject<Block> COPPER_BLOCK = BlocksHolder.registerBlock(
+    public static final RegistryObject<Block> COPPER_BLOCK = OIBlocks.registerBlock(
             "copper_block",
             () -> new Block(
                     BlockBehaviour.Properties
@@ -56,45 +59,73 @@ public final class BlocksHolder {
 
     //endregion
 
+    // region SILVER
+    public static final RegistryObject<Block> SILVER_ORE = OIBlocks.registerBlock(
+            "silver_ore",
+            () -> new Block(
+                    BlockBehaviour.Properties
+                            .of(Material.STONE)
+                            .strength(3.5f)
+                            .requiresCorrectToolForDrops()
+            ),
+            OI_TAB
+    );
+
+    public static final RegistryObject<Block> SILVER_BLOCK = OIBlocks.registerBlock(
+            "silver_block",
+            () -> new Block(
+                    BlockBehaviour.Properties
+                            .of(Material.METAL)
+                            .strength(5.5f)
+            ),
+            OI_TAB
+    );
+
+    //endregion
+
     //region RUBBER TREE
 
-    public static final RegistryObject<Block> RUBBER_LOG = registerBlock(
+    public static final RegistryObject<Block> RUBBER_LOG = registerFuelBlock(
             "rubber_log",
             () -> new OIFlammableRotatedPillarBlock(
                     BlockBehaviour.Properties.copy(OAK_LOG)
                             .requiresCorrectToolForDrops()
             ),
-            OI_TAB
+            OI_TAB,
+            300
     );
 
-    public static final RegistryObject<Block> RUBBER_WOOD = registerBlock(
+    public static final RegistryObject<Block> RUBBER_WOOD = registerFuelBlock(
             "rubber_wood",
             () -> new OIFlammableRotatedPillarBlock(
                     BlockBehaviour.Properties.copy(OAK_WOOD)
                             .requiresCorrectToolForDrops()
             ),
-            OI_TAB
+            OI_TAB,
+            300
     );
 
-    public static final RegistryObject<Block> STRIPPED_RUBBER_LOG = registerBlock(
+    public static final RegistryObject<Block> STRIPPED_RUBBER_LOG = registerFuelBlock(
             "stripped_rubber_log",
             () -> new OIFlammableRotatedPillarBlock(
                     BlockBehaviour.Properties.copy(STRIPPED_OAK_LOG)
                             .requiresCorrectToolForDrops()
             ),
-            OI_TAB
+            OI_TAB,
+            300
     );
 
-    public static final RegistryObject<Block> STRIPPED_RUBBER_WOOD = registerBlock(
+    public static final RegistryObject<Block> STRIPPED_RUBBER_WOOD = registerFuelBlock(
             "stripped_rubber_wood",
             () -> new OIFlammableRotatedPillarBlock(
                     BlockBehaviour.Properties.copy(STRIPPED_OAK_WOOD)
                             .requiresCorrectToolForDrops()
             ),
-            OI_TAB
+            OI_TAB,
+            300
     );
 
-    public static final RegistryObject<Block> RUBBER_PLANKS = registerBlock(
+    public static final RegistryObject<Block> RUBBER_PLANKS = registerFuelBlock(
             "rubber_planks",
             () -> new Block(
                     BlockBehaviour.Properties.copy(OAK_PLANKS)
@@ -114,7 +145,8 @@ public final class BlocksHolder {
                     return 20;
                 }
             },
-            OI_TAB
+            OI_TAB,
+            300
     );
 
     public static final RegistryObject<Block> RUBBER_LEAVES = registerBlock(
@@ -141,13 +173,14 @@ public final class BlocksHolder {
             OI_TAB
     );
 
-    public static final RegistryObject<Block> RUBBER_SAPLING = registerBlock(
+    public static final RegistryObject<Block> RUBBER_SAPLING = registerFuelBlock(
             "rubber_sapling",
             () -> new SaplingBlock(
                     new RubberTreeGrower(),
                     BlockBehaviour.Properties.copy(OAK_SAPLING)
             ),
-            OI_TAB
+            OI_TAB,
+            100
     );
 
 
@@ -165,6 +198,21 @@ public final class BlocksHolder {
     }
 
     public static <T extends Block> void registerBlockItem(String name, Supplier<T> block, CreativeModeTab tab) {
-        ItemsHolder.ITEMS.register(name, () -> new BlockItem(block.get(), new Item.Properties().tab(tab)));
+        OIItems.ITEMS.register(name, () -> new BlockItem(block.get(), new Item.Properties().tab(tab)));
+    }
+
+    public static <T extends Block> RegistryObject<T> registerFuelBlock(String name, Supplier<T> block, CreativeModeTab tab, int burnTime) {
+        RegistryObject<T> toReturn = BLOCKS.register(name, block);
+        registerFuelBlockItem(name, toReturn, tab, burnTime);
+        return toReturn;
+    }
+
+    public static <T extends Block> void registerFuelBlockItem(String name, Supplier<T> block, CreativeModeTab tab, int burnTime) {
+        OIItems.ITEMS.register(name, () -> new BlockItem(block.get(), new Item.Properties().tab(tab)){
+            @Override
+            public int getBurnTime(ItemStack itemStack, @Nullable RecipeType<?> recipeType) {
+                return burnTime;
+            }
+        });
     }
 }