+++ src/main/java/com/angelobdev/openindustry/world/feature/OIConfiguredFeatures.java
@@ -1,7 +1,7 @@
 package com.angelobdev.openindustry.world.feature;
 
 import com.angelobdev.openindustry.OpenIndustry;
-import com.angelobdev.openindustry.block.BlocksHolder;
+import com.angelobdev.openindustry.block.OIBlocks;
 import com.google.common.base.Suppliers;
 import net.minecraft.core.Registry;
 import net.minecraft.data.worldgen.features.OreFeatures;
@@ -17,6 +17,7 @@ import net.minecraft.world.level.levelgen.feature.trunkplacers.StraightTrunkPlac
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.RegistryObject;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.function.Supplier;
@@ -26,15 +27,29 @@ public class OIConfiguredFeatures {
     public static final DeferredRegister<ConfiguredFeature<?, ?>> CONFIGURED_FEATURES = DeferredRegister.create(Registry.CONFIGURED_FEATURE_REGISTRY, OpenIndustry.MODID);
 
     //region COPPER ORE
-    public static final Supplier<List<OreConfiguration.TargetBlockState>> OVERWORLD_COPPER_ORE = Suppliers.memoize(
+    public static final @NotNull Supplier<List<OreConfiguration.TargetBlockState>> OVERWORLD_COPPER_ORE = Suppliers.memoize(
             () -> List.of(
-                    OreConfiguration.target(OreFeatures.STONE_ORE_REPLACEABLES, BlocksHolder.COPPER_ORE.get().defaultBlockState())
+                    OreConfiguration.target(OreFeatures.STONE_ORE_REPLACEABLES, OIBlocks.COPPER_ORE.get().defaultBlockState())
             )
     );
 
-    public static final RegistryObject<ConfiguredFeature<?, ?>> COPPER_ORE = CONFIGURED_FEATURES.register(
+    public static final @NotNull RegistryObject<ConfiguredFeature<?, ?>> COPPER_ORE = CONFIGURED_FEATURES.register(
             "copper_ore",
             () -> new ConfiguredFeature<>(Feature.ORE, new OreConfiguration(OVERWORLD_COPPER_ORE.get(), 7)) // 7 is the vein-size
+    );
+    //endregion
+
+    //region SILVER ORE
+
+    public static final Supplier<List<OreConfiguration.TargetBlockState>> OVERWORLD_SILVER_ORE = Suppliers.memoize(
+            () -> List.of(
+                    OreConfiguration.target(OreFeatures.STONE_ORE_REPLACEABLES, OIBlocks.SILVER_ORE.get().defaultBlockState())
+            )
+    );
+
+    public static final RegistryObject<ConfiguredFeature<?, ?>> SILVER_ORE = CONFIGURED_FEATURES.register(
+            "silver_ore",
+            () -> new ConfiguredFeature<>(Feature.ORE, new OreConfiguration(OVERWORLD_SILVER_ORE.get(), 7)) // 7 is the vein-size
     );
     //endregion
 
@@ -45,9 +60,9 @@ public class OIConfiguredFeatures {
             () -> new ConfiguredFeature<>(
                     Feature.TREE,
                     new TreeConfiguration.TreeConfigurationBuilder(
-                            BlockStateProvider.simple(BlocksHolder.RUBBER_LOG.get()),
+                            BlockStateProvider.simple(OIBlocks.RUBBER_LOG.get()),
                             new StraightTrunkPlacer(5, 6, 3),
-                            BlockStateProvider.simple(BlocksHolder.RUBBER_LEAVES.get()),
+                            BlockStateProvider.simple(OIBlocks.RUBBER_LEAVES.get()),
                             new BlobFoliagePlacer(ConstantInt.of(2), ConstantInt.of(0), 4),
                             new TwoLayersFeatureSize(1, 0, 2)
                     ).build()